+++ pkg/helpers/strings.go
@@ -1,0 +1,14 @@
+package helpers
+
+import "strings"
+
+func CleanString(s string) string {
+	s = strings.ReplaceAll(s, "[", "")
+	s = strings.ReplaceAll(s, "]", "")
+	s = strings.ReplaceAll(s, "{", "")
+	s = strings.ReplaceAll(s, "}", "")
+	s = strings.ReplaceAll(s, "(", "")
+	s = strings.ReplaceAll(s, ")", "")
+
+	return s
+}